Create a shared structure.
